dev_data.yaml: fix test data fixtures ordering
When spinning up a test instance via ``docker-compose up``, Postgres is
used as the default database. Postgres enforces foreign key constraints
on insert and so attempting to populate the database with test data
from ``dev_data.yaml`` will fail. This is because test fixtures for
Event-s with primary keys 6 and 7 reference a non-NULL event group
before those have been added to the DB.

Re-order the test fixture so that it can be loaded on DBs which enforce
foreign key constraints by simply adding the EvenGroup models first.

It may be beneficial to enable foreign key constraint checking in sqlite
going-forward so that this can be caught automatically in testing. See,
for example,

  https://stackoverflow.com/questions/8042845/django-how-to-enable-foreign-keys-in-sqlite3-backend