dev_data.yaml: shorten slug
The default maximum length of a slig field in Djano is 50 characters.
The example data fixture has a slug which is longer than that. Modify
the slug to fit within the maximum length. This is triggered when
running with the default docker-compose setup since Postgres enforces
the length constraint.

Unfortunately there is no PRAGMA to set when using sqlite to make sure
this can be caught in testing.

An alternative might be to increase the length of the slug field and
create a new migration.