Minor tweaks to help with collection administration
The owner of a public collection is not listed in the view - this causes confusion if people want to contribute. Changing ownership of a collection is almost impossible with a massive unordered drop-down list of users so I've deployed the raw-field option. This means looking up the ID of the user but that's OK as I have to find their username anyway.