Hack for deleted departments
Architecturally it would be a nightmare to attempt to fix this as this code can't know whether departments are deleted before requesting them. The code that makes the request raises the response as an exception if it is an error response (in this case 400 bad request) - as it should. So that shouldn't be changed either. Hence this hack, and we will not delete oxpoints departments in the future.